+++ spec/spec_helper.rb
@@ -24,6 +24,7 @@ UNIQUE_TEST_ID = "#{TEST_TAG}_#{SUFFIX}"
 UNIQUE_TEST_FOLDER = "#{TEST_TAG}_#{SUFFIX}_folder"
 NEXT_CURSOR = "db27cfb02b3f69cb39049969c23ca430c6d33d5a3a7c3ad1d870c54e1a54ee0faa5acdd9f6d288666986001711759d10"
 GENERIC_FOLDER_NAME = "some_folder"
+UPLOADER_TAG = "#{TEST_TAG}_uploader"
 
 EVAL_STR='if (resource_info["width"] < 450) { upload_options["quality_analysis"] = true };
           upload_options["context"] = "width=" + resource_info["width"]'
@@ -72,10 +73,44 @@ RSpec.configure do |config|
   end
 end
 
+RSpec.shared_context "config" do
+  before do
+    Cloudinary.reset_config
+    @cloudinary_url_backup = ENV["CLOUDINARY_URL"]
+    @account_url_backup = ENV["CLOUDINARY_ACCOUNT_URL"]
+  end
+
+  after do
+    ENV.keys.select { |key| key.start_with? "CLOUDINARY_" }.each { |key| ENV.delete(key) }
+    ENV["CLOUDINARY_ACCOUNT_URL"] = @account_url_backup
+    ENV["CLOUDINARY_URL"] = @cloudinary_url_backup
+    Cloudinary.reset_config
+  end
+end
+
 RSpec.shared_context "cleanup" do |tag|
   tag ||= TEST_TAG
   after :all do
     Cloudinary::Api.delete_resources_by_tag(tag) unless Cloudinary.config.keep_test_products
+  end
+end
+
+RSpec.shared_context "metadata_field" do |metadata_attributes|
+  metadata_external_ids = []
+
+  before(:all) do
+    metadata_attributes = metadata_attributes.is_a?(Array) ? metadata_attributes : [metadata_attributes]
+
+    metadata_attributes.each do |attributes|
+      metadata_external_ids << attributes[:external_id]
+      Cloudinary::Api.add_metadata_field(attributes)
+    end
+  end
+
+  after(:all) do
+    metadata_external_ids.each do |metadata_external_id|
+      Cloudinary::Api.delete_metadata_field(metadata_external_id)
+    end
   end
 end